+++ clients/admin-ui/cypress/e2e/data-catalog.cy.ts
@@ -29,15 +29,6 @@ describe("data catalog", () => {
         "BigQuery System",
       );
     });
-
-    it("should be able to navigate to system details via the overflow menu", () => {
-      cy.getByTestId("row-bigquery_system").within(() => {
-        cy.getByTestId("system-actions-menu").click();
-        cy.getByTestId("view-system-details").click({ force: true });
-        cy.url().should("include", "/systems/configure/bigquery_system");
-      });
-    });
-
     it("should be able to add a data use", () => {
       cy.getByTestId("row-bigquery_system-col-data-uses").within(() => {
         cy.getByTestId("taxonomy-add-btn").click();
@@ -58,6 +49,29 @@ describe("data catalog", () => {
       cy.getByTestId("row-dynamo_system-col-name").click();
       cy.wait("@getEmptyAvailableDatabases");
       cy.url().should("not.include", "/projects");
+    });
+
+    describe("system detail drawer", () => {
+      it("should be able to navigate to system details via the detail drawer", () => {
+        cy.getByTestId("row-bigquery_system").within(() => {
+          cy.getByTestId("system-actions-menu").click();
+          cy.getByTestId("view-system-details").click({ force: true });
+        });
+        cy.getByTestId("system-details").should("be.visible");
+        cy.getByTestId("edit-system-btn").click();
+        cy.url().should("include", "/systems/configure/bigquery_system");
+      });
+
+      it("should be able to add data uses via the detail drawer", () => {
+        cy.getByTestId("row-bigquery_system").within(() => {
+          cy.getByTestId("system-actions-menu").click();
+          cy.getByTestId("view-system-details").click({ force: true });
+        });
+        cy.getByTestId("system-details").within(() => {
+          cy.getByTestId("taxonomy-add-btn").click();
+          cy.get(".select-wrapper").should("be.visible");
+        });
+      });
     });
   });
 
@@ -91,13 +105,23 @@ describe("data catalog", () => {
         "/projects/bigquery_monitor.prj-bigquery-111111",
       );
     });
+
+    it("should be able to view details in the drawer", () => {
+      cy.getByTestId(
+        "row-bigquery_monitor.prj-bigquery-111111-col-actions",
+      ).within(() => {
+        cy.getByTestId("actions-menu").click();
+        cy.getByTestId("view-resource-details").click({ force: true });
+      });
+      cy.getByTestId("resource-details").should("be.visible");
+    });
   });
 
   describe("resource tables", () => {
     beforeEach(() => {
       stubStagedResourceActions();
       cy.visit(
-        `${DATA_CATALOG_ROUTE}/bigquery_system/monitor.project.test_dataset_1`,
+        `${DATA_CATALOG_ROUTE}/bigquery_system/projects/monitor.project/monitor.project.test_dataset_1`,
       );
     });
 
@@ -117,8 +141,8 @@ describe("data catalog", () => {
       );
       cy.getByTestId("row-monitor.project.dataset.table_2-col-actions").within(
         () => {
-          cy.getByTestId("resource-actions-menu").click();
-          cy.getByTestId("hide-action").click({ force: true });
+          cy.getByTestId("actions-menu").click();
+          cy.getByTestId("hide-resource").click({ force: true });
           cy.wait("@ignoreResource");
         },
       );
@@ -128,6 +152,30 @@ describe("data catalog", () => {
           cy.wait("@promoteResource");
         },
       );
+    });
+
+    describe("resource detail drawer", () => {
+      it("should be able to view resource details", () => {
+        cy.getByTestId("row-monitor.project.dataset.table_1").within(() => {
+          cy.getByTestId("actions-menu").click();
+          cy.getByTestId("view-resource-details").click({ force: true });
+        });
+        cy.getByTestId("resource-details")
+          .should("be.visible")
+          .within(() => {
+            // don't edit categories unless resource has the right status
+            cy.getByTestId("edit-category-cell").should("not.exist");
+          });
+      });
+      it("should be able to edit data categories where applicable", () => {
+        cy.getByTestId("row-monitor.project.dataset.table_3").within(() => {
+          cy.getByTestId("actions-menu").click();
+          cy.getByTestId("view-resource-details").click({ force: true });
+        });
+        cy.getByTestId("resource-details").within(() => {
+          cy.getByTestId("user-classification-system").should("exist");
+        });
+      });
     });
   });
 });