+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/hooks/useDiscoveredSystemAggregateColumns.tsx
@@ -1,6 +1,10 @@
 import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 
-import { DefaultCell } from "~/features/common/table/v2";
+import {
+  DefaultCell,
+  IndeterminateCheckboxCell,
+} from "~/features/common/table/v2";
+import DiscoveredSystemDataUseCell from "~/features/data-discovery-and-detection/action-center/tables/cells/DiscoveredSystemDataUseCell";
 
 import { DiscoveredSystemActionsCell } from "../tables/cells/DiscoveredSystemAggregateActionsCell";
 import { DiscoveredSystemStatusCell } from "../tables/cells/DiscoveredSystemAggregateStatusCell";
@@ -10,6 +14,28 @@ export const useDiscoveredSystemAggregateColumns = (monitorId: string) => {
   const columnHelper = createColumnHelper<MonitorSystemAggregate>();
 
   const columns: ColumnDef<MonitorSystemAggregate, any>[] = [
+    columnHelper.display({
+      id: "select",
+      cell: ({ row }) => (
+        <IndeterminateCheckboxCell
+          isChecked={row.getIsSelected()}
+          onChange={row.getToggleSelectedHandler()}
+          dataTestId={`select-${row.original.name || row.id}`}
+        />
+      ),
+      header: ({ table }) => (
+        <IndeterminateCheckboxCell
+          isChecked={table.getIsAllPageRowsSelected()}
+          isIndeterminate={table.getIsSomeRowsSelected()}
+          onChange={table.getToggleAllRowsSelectedHandler()}
+          dataTestId="select-all-rows"
+        />
+      ),
+      maxSize: 40,
+      meta: {
+        disableRowClick: true,
+      },
+    }),
     columnHelper.accessor((row) => row.name, {
       id: "system_name",
       cell: (props) => (
@@ -26,15 +52,17 @@ export const useDiscoveredSystemAggregateColumns = (monitorId: string) => {
       header: "Assets",
       size: 80,
     }),
-    /*
-    // TODO: [HJ-356] uncomment when monitor supports categories of consent
     columnHelper.display({
       id: "data_use",
+      cell: (props) => (
+        <DiscoveredSystemDataUseCell system={props.row.original} />
+      ),
       header: "Categories of consent",
       meta: {
         width: "auto",
+        disableRowClick: true,
       },
-    }), */
+    }),
     columnHelper.accessor((row) => row.locations, {
       id: "locations",
       cell: (props) => (