+++ clients/admin-ui/src/pages/data-catalog/[systemId]/projects/[projectUrn]/index.tsx
@@ -4,8 +4,9 @@ import {
   getGroupedRowModel,
   useReactTable,
 } from "@tanstack/react-table";
+import { Icons } from "fidesui";
 import { useRouter } from "next/router";
-import { useEffect, useMemo } from "react";
+import { useEffect, useMemo, useState } from "react";
 
 import Layout from "~/features/common/Layout";
 import { DATA_CATALOG_ROUTE } from "~/features/common/nav/v2/routes";
@@ -16,9 +17,11 @@ import {
   TableSkeletonLoader,
   useServerSidePagination,
 } from "~/features/common/table/v2";
-import { useGetCatalogDatasetsQuery } from "~/features/data-catalog/data-catalog.slice";
+import CatalogDatasetDetailDrawer from "~/features/data-catalog/datasets/CatalogDatasetDetailDrawer";
 import EmptyCatalogTableNotice from "~/features/data-catalog/datasets/EmptyCatalogTableNotice";
 import useCatalogDatasetColumns from "~/features/data-catalog/datasets/useCatalogDatasetColumns";
+import { getProjectName } from "~/features/data-catalog/utils/urnParsing";
+import { useGetMonitorResultsQuery } from "~/features/data-discovery-and-detection/discovery-detection.slice";
 import { useGetSystemByFidesKeyQuery } from "~/features/system";
 import { StagedResourceAPIResponse } from "~/types/api";
 
@@ -30,10 +33,14 @@ const EMPTY_RESPONSE = {
   pages: 1,
 };
 
-const CatalogDatasetViewNoProjects = () => {
+const CatalogDatasetView = () => {
   const { query, push } = useRouter();
   const systemKey = query.systemId as string;
-  const monitorConfigKeys = query.monitor_config_ids as string[];
+  const projectUrn = query.projectUrn as string;
+
+  const [detailsToView, setDetailsToView] = useState<
+    StagedResourceAPIResponse | undefined
+  >(undefined);
 
   const { data: system, isLoading: systemIsLoading } =
     useGetSystemByFidesKeyQuery(systemKey);
@@ -56,10 +63,10 @@ const CatalogDatasetViewNoProjects = () => {
     isFetching,
     isLoading,
     data: resources,
-  } = useGetCatalogDatasetsQuery({
+  } = useGetMonitorResultsQuery({
+    staged_resource_urn: projectUrn,
     page: pageIndex,
     size: pageSize,
-    monitor_config_ids: monitorConfigKeys,
   });
 
   const {
@@ -72,7 +79,9 @@ const CatalogDatasetViewNoProjects = () => {
     setTotalPages(totalPages);
   }, [totalPages, setTotalPages]);
 
-  const columns = useCatalogDatasetColumns();
+  const columns = useCatalogDatasetColumns({
+    onDetailClick: (dataset) => setDetailsToView(dataset),
+  });
 
   const tableInstance = useReactTable<StagedResourceAPIResponse>({
     getCoreRowModel: getCoreRowModel(),
@@ -94,7 +103,9 @@ const CatalogDatasetViewNoProjects = () => {
           { title: "All systems", href: DATA_CATALOG_ROUTE },
           {
             title: system?.name || systemKey,
+            href: DATA_CATALOG_ROUTE,
           },
+          { title: getProjectName(projectUrn), icon: <Icons.Db2Database /> },
         ]}
       />
       {!showContent && <TableSkeletonLoader rowHeight={36} numRows={36} />}
@@ -104,7 +115,9 @@ const CatalogDatasetViewNoProjects = () => {
             tableInstance={tableInstance}
             emptyTableNotice={<EmptyCatalogTableNotice />}
             onRowClick={(row) =>
-              push(`${DATA_CATALOG_ROUTE}/${systemKey}/${row.urn}`)
+              push(
+                `${DATA_CATALOG_ROUTE}/${systemKey}/projects/${projectUrn}/${row.urn}/`,
+              )
             }
           />
           <PaginationBar
@@ -118,10 +131,14 @@ const CatalogDatasetViewNoProjects = () => {
             startRange={startRange}
             endRange={endRange}
           />
+          <CatalogDatasetDetailDrawer
+            dataset={detailsToView}
+            onClose={() => setDetailsToView(undefined)}
+          />
         </>
       )}
     </Layout>
   );
 };
 
-export default CatalogDatasetViewNoProjects;
+export default CatalogDatasetView;