+++ clients/fides-js/src/components/tcf/TcfOverlay.tsx
@@ -9,9 +9,14 @@ import {
   ConsentMethod,
   PrivacyExperience,
   PrivacyExperienceMinimal,
+  PrivacyNoticeWithPreference,
+  SaveConsentPreference,
   ServingComponent,
 } from "../../lib/consent-types";
-import { isPrivacyExperience } from "../../lib/consent-utils";
+import {
+  experienceIsValid,
+  isPrivacyExperience,
+} from "../../lib/consent-utils";
 import { dispatchFidesEvent } from "../../lib/events";
 import { useNoticesServed } from "../../lib/hooks";
 import {
@@ -22,13 +27,16 @@ import {
   loadMessagesFromGVLTranslations,
   matchAvailableLocales,
   selectBestExperienceConfigTranslation,
+  selectBestNoticeTranslation,
 } from "../../lib/i18n";
 import { useI18n } from "../../lib/i18n/i18n-context";
 import { updateConsentPreferences } from "../../lib/preferences";
+import { transformConsentToFidesUserPreference } from "../../lib/shared-consent-utils";
 import { EMPTY_ENABLED_IDS } from "../../lib/tcf/constants";
 import { useGvl } from "../../lib/tcf/gvl-context";
-import type {
+import {
   EnabledIds,
+  PrivacyNoticeWithBestTranslation,
   TcfModels,
   TcfSavePreferences,
 } from "../../lib/tcf/types";
@@ -38,6 +46,7 @@ import {
   createTcfSavePayload,
   createTcfSavePayloadFromMinExp,
   getEnabledIds,
+  getEnabledIdsNotice,
   getGVLPurposeList,
   updateCookie,
 } from "../../lib/tcf/utils";
@@ -51,7 +60,9 @@ import { TCFBannerSupplemental } from "./TCFBannerSupplemental";
 import { TcfConsentButtons } from "./TcfConsentButtons";
 import TcfTabs from "./TcfTabs";
 
-const getAllIds = (modelList: TcfModels) => {
+const getAllIds = (
+  modelList: TcfModels | Array<PrivacyNoticeWithPreference>,
+) => {
   if (!modelList) {
     return [];
   }
@@ -61,6 +72,7 @@ const getAllIds = (modelList: TcfModels) => {
 interface TcfOverlayProps extends Omit<OverlayProps, "experience"> {
   experienceMinimal: PrivacyExperienceMinimal;
 }
+
 export const TcfOverlay = ({
   options,
   experienceMinimal,
@@ -111,6 +123,24 @@ export const TcfOverlay = ({
   };
 
   /**
+   * Enhance the given PrivacyNotices with best translation for rendering.
+   *
+   * We memoize these together to avoid repeatedly figuring out the "best"
+   * translation on every render, since it will only change if the overall
+   * locale changes!
+   */
+  const privacyNoticesWithBestTranslation: PrivacyNoticeWithBestTranslation[] =
+    useMemo(
+      () =>
+        (experienceMinimal.privacy_notices || []).map((notice) => {
+          const bestTranslation = selectBestNoticeTranslation(i18n, notice);
+          return { ...notice, bestTranslation };
+        }),
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+      [experienceMinimal.privacy_notices, i18n, currentLocale],
+    );
+
+  /**
    * TCF overlay loads with a minimal experience, which is then replaced with the full.
    * We do this to ensure the overlay can be rendered as quickly as possible.
    * The full experience is fetched after the component mounts, so we store it
@@ -141,6 +171,7 @@ export const TcfOverlay = ({
         isGVLLangLoading = false;
       });
     }
+    fidesDebugger("Fetching full TCF experience...");
     fetchExperience({
       userLocationString: fidesRegionString,
       fidesApiUrl: options.fidesApiUrl,
@@ -151,17 +182,29 @@ export const TcfOverlay = ({
       if (isPrivacyExperience(result)) {
         // include user preferences from the cookie
         const userPrefs = buildUserPrefs(result, cookie);
-        const fullExperience: PrivacyExperience = { ...result, ...userPrefs };
 
-        setExperience(fullExperience);
-        loadMessagesFromExperience(i18n, fullExperience, translationOverrides);
-        if (!userlocale || bestLocale === defaultLocale) {
-          // English (default) GVL translations are part of the full experience, so we load them here.
-          loadGVLMessagesFromExperience(i18n, fullExperience);
-        } else {
-          setCurrentLocale(bestLocale);
-          if (!isGVLLangLoading) {
-            setIsI18nLoading(false);
+        if (experienceIsValid(result)) {
+          const fullExperience: PrivacyExperience = { ...result, ...userPrefs };
+          window.Fides.experience = {
+            ...window.Fides.experience,
+            ...fullExperience,
+          };
+          window.Fides.experience.minimal_tcf = false;
+
+          setExperience(fullExperience);
+          loadMessagesFromExperience(
+            i18n,
+            fullExperience,
+            translationOverrides,
+          );
+          if (!userlocale || bestLocale === defaultLocale) {
+            // English (default) GVL translations are part of the full experience, so we load them here.
+            loadGVLMessagesFromExperience(i18n, fullExperience);
+          } else {
+            setCurrentLocale(bestLocale);
+            if (!isGVLLangLoading) {
+              setIsI18nLoading(false);
+            }
           }
         }
       }
@@ -179,6 +222,7 @@ export const TcfOverlay = ({
     } else {
       const {
         tcf_purpose_consents: consentPurposes = [],
+        privacy_notices: customPurposes = [],
         tcf_purpose_legitimate_interests: legintPurposes = [],
         tcf_special_purposes: specialPurposes = [],
         tcf_features: features = [],
@@ -192,6 +236,7 @@ export const TcfOverlay = ({
       // Vendors and systems are the same to the FE, so we combine them here
       setDraftIds({
         purposesConsent: getEnabledIds(consentPurposes),
+        customPurposesConsent: getEnabledIdsNotice(customPurposes),
         purposesLegint: getEnabledIds(legintPurposes),
         specialPurposes: getEnabledIds(specialPurposes),
         features: getEnabledIds(features),
@@ -223,6 +268,10 @@ export const TcfOverlay = ({
     return undefined;
   }, [experienceMinimal, experience, i18n]);
 
+  const customPurposes: (string | undefined)[] = useMemo(() => {
+    return privacyNoticesWithBestTranslation.map((notice) => notice.name);
+  }, [privacyNoticesWithBestTranslation]);
+
   const purposes: string[] = useMemo(() => {
     if (gvlTranslations) {
       return getGVLPurposeList(gvlTranslations);
@@ -240,13 +289,42 @@ export const TcfOverlay = ({
 
   const { servedNoticeHistoryId } = useNoticesServed({
     privacyExperienceConfigHistoryId,
-    privacyNoticeHistoryIds: [],
+    privacyNoticeHistoryIds: privacyNoticesWithBestTranslation.reduce(
+      (ids, e) => {
+        const id = e.bestTranslation?.privacy_notice_history_id;
+        if (id) {
+          ids.push(id);
+        }
+        return ids;
+      },
+      [] as string[],
+    ),
     options,
     userGeography: fidesRegionString,
     acknowledgeMode: false,
     privacyExperience: experience || experienceMinimal,
     tcfNoticesServed,
   });
+
+  const createConsentPreferencesToSave = (
+    privacyNoticeList: PrivacyNoticeWithBestTranslation[],
+    enabledPrivacyNoticeIds: string[],
+  ): SaveConsentPreference[] => {
+    if (!privacyNoticeList || !enabledPrivacyNoticeIds) {
+      return [];
+    }
+    return privacyNoticeList.map((item) => {
+      const userPreference = transformConsentToFidesUserPreference(
+        enabledPrivacyNoticeIds.includes(item.id),
+        item.consent_mechanism,
+      );
+      return new SaveConsentPreference(
+        item,
+        userPreference,
+        item.bestTranslation?.privacy_notice_history_id,
+      );
+    });
+  };
 
   const handleUpdateAllPreferences = useCallback(
     (consentMethod: ConsentMethod, enabledIds: EnabledIds) => {
@@ -265,8 +343,12 @@ export const TcfOverlay = ({
           enabledIds,
         });
       }
+      const consentPreferencesToSave = createConsentPreferencesToSave(
+        privacyNoticesWithBestTranslation,
+        enabledIds.customPurposesConsent,
+      );
       updateConsentPreferences({
-        consentPreferencesToSave: [],
+        consentPreferencesToSave,
         privacyExperienceConfigHistoryId,
         experience: experience || experienceMinimal,
         consentMethod,
@@ -282,6 +364,7 @@ export const TcfOverlay = ({
             tcf,
             enabledIds,
             experience || experienceMinimal,
+            consentPreferencesToSave,
           ),
       });
       setDraftIds(enabledIds);
@@ -293,6 +376,7 @@ export const TcfOverlay = ({
       fidesRegionString,
       options,
       privacyExperienceConfigHistoryId,
+      privacyNoticesWithBestTranslation,
       servedNoticeHistoryId,
     ],
   );
@@ -305,6 +389,7 @@ export const TcfOverlay = ({
         exp = experience as PrivacyExperience;
         allIds = {
           purposesConsent: getAllIds(exp.tcf_purpose_consents),
+          customPurposesConsent: getAllIds(exp.privacy_notices),
           purposesLegint: getAllIds(exp.tcf_purpose_legitimate_interests),
           specialPurposes: getAllIds(exp.tcf_special_purposes),
           features: getAllIds(exp.tcf_features),
@@ -324,6 +409,7 @@ export const TcfOverlay = ({
         allIds = {
           purposesConsent:
             exp.tcf_purpose_consent_ids?.map((id) => `${id}`) || [],
+          customPurposesConsent: getAllIds(exp.privacy_notices) || [],
           purposesLegint:
             exp.tcf_purpose_legitimate_interest_ids?.map((id) => `${id}`) || [],
           specialPurposes:
@@ -451,7 +537,10 @@ export const TcfOverlay = ({
             )}
             className="fides-tcf-banner-container"
           >
-            <TCFBannerSupplemental purposes={purposes} />
+            <TCFBannerSupplemental
+              purposes={purposes}
+              customPurposes={customPurposes}
+            />
           </ConsentBanner>
         );
       }}
@@ -461,6 +550,7 @@ export const TcfOverlay = ({
           : () => (
               <TcfTabs
                 experience={experience}
+                customNotices={privacyNoticesWithBestTranslation}
                 enabledIds={draftIds}
                 onChange={(updatedIds) => {
                   setDraftIds(updatedIds);