+++ clients/fides-js/src/components/tcf/TcfTabs.tsx
@@ -3,7 +3,10 @@ import { useCallback, useRef } from "preact/hooks";
 
 import { PrivacyExperience } from "../../lib/consent-types";
 import { useI18n } from "../../lib/i18n/i18n-context";
-import { EnabledIds } from "../../lib/tcf/types";
+import {
+  EnabledIds,
+  PrivacyNoticeWithBestTranslation,
+} from "../../lib/tcf/types";
 import InfoBox from "../InfoBox";
 import TcfFeatures from "./TcfFeatures";
 import TcfPurposes from "./TcfPurposes";
@@ -19,12 +22,14 @@ export interface UpdateEnabledIds {
 
 const TcfTabs = ({
   experience,
+  customNotices,
   enabledIds,
   onChange,
   activeTabIndex,
   onTabChange,
 }: {
   experience: PrivacyExperience;
+  customNotices: PrivacyNoticeWithBestTranslation[] | undefined;
   enabledIds: EnabledIds;
   onChange: (payload: EnabledIds) => void;
   activeTabIndex: number;
@@ -48,9 +53,11 @@ const TcfTabs = ({
           <InfoBox>{i18n.t("static.tcf.purposes.description")}</InfoBox>
           <TcfPurposes
             allPurposesConsent={experience.tcf_purpose_consents}
+            allCustomPurposesConsent={customNotices}
             allPurposesLegint={experience.tcf_purpose_legitimate_interests}
             allSpecialPurposes={experience.tcf_special_purposes}
             enabledPurposeConsentIds={enabledIds.purposesConsent}
+            enabledCustomPurposeConsentIds={enabledIds.customPurposesConsent}
             enabledPurposeLegintIds={enabledIds.purposesLegint}
             enabledSpecialPurposeIds={enabledIds.specialPurposes}
             onChange={handleUpdateDraftState}